refactor(UI): settings page (#165)
* refactor: settings page

* i18n

* refactor

* refactor

* refactor

* fix: i18n
