+++ lib/router/router.dart
@@ -14,7 +14,7 @@ import 'package:miru_app/views/pages/home_page.dart';
 import 'package:miru_app/views/pages/main_page.dart';
 import 'package:miru_app/views/pages/search/extension_searcher_page.dart';
 import 'package:miru_app/views/pages/search/search_page.dart';
-import 'package:miru_app/views/pages/settings_page.dart';
+import 'package:miru_app/views/pages/settings/settings_page.dart';
 
 final rootNavigatorKey = GlobalKey<NavigatorState>();
 final _shellNavigatorKey = GlobalKey<NavigatorState>();
