+++ lib/views/widgets/settings/settings_switch_tile.dart
@@ -1,6 +1,6 @@
 import 'package:fluent_ui/fluent_ui.dart' as fluent;
 import 'package:flutter/material.dart';
-import 'package:miru_app/views/widgets/settings_tile.dart';
+import 'package:miru_app/views/widgets/settings/settings_tile.dart';
 import 'package:miru_app/views/widgets/platform_widget.dart';
 
 class SettingsSwitchTile extends StatefulWidget {
@@ -11,12 +11,14 @@ class SettingsSwitchTile extends StatefulWidget {
     required this.buildValue,
     required this.onChanged,
     this.buildSubtitle,
+    this.isCard = false,
   });
   final Widget? icon;
   final String title;
   final String Function()? buildSubtitle;
   final bool Function() buildValue;
   final Function(bool) onChanged;
+  final bool isCard;
 
   @override
   State<SettingsSwitchTile> createState() => _SettingsSwitchTileState();
@@ -26,6 +28,7 @@ class _SettingsSwitchTileState extends State<SettingsSwitchTile> {
   @override
   Widget build(BuildContext context) {
     return SettingsTile(
+      isCard: widget.isCard,
       icon: widget.icon,
       title: widget.title,
       buildSubtitle: widget.buildSubtitle,
